+++ docs/CHANGELOG.md
@@ -1466,10 +1466,12 @@ for OCR jobs.
 
 - Brings `/runs` tab back to the operator UI.
 - Signs out a user from operator UI on authentication error.
+- OCR jobs no longer require defining v1 bootstrap peers unless `P2P_NETWORKING_STACK=V1`
 
 #### BREAKING CHANGES
 
 - Commands for creating/managing legacy jobs and OCR jobs have changed, to reduce confusion and accommodate additional types of jobs using the new pipeline.
+- If `P2P_NETWORKING_STACK=V1V2`, then `P2P_BOOTSTRAP_PEERS` must also be set
 
 #### V1 jobs